weather_fetcher.py
import requests, os
from dotenv import load_dotenv

load_dotenv()
API_KEY = os.getenv("OPENWEATHERMAP_API_KEY")

def get_weather(city):
    url = f"http://api.openweathermap.org/data/2.5/weather?q={city}&appid={API_KEY}&units=metric"
    res = requests.get(url).json()
    weather = res["weather"][0]["main"]
    temp = res["main"]["temp"]
    return weather, temp