requirements.txt
streamlit
opencv-python
requests
spotipy
python-dotenv